+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.84.0]
 
 * Improve the "Convert module to object" refactoring to handle multi-line module export lists with interspersed comments
+* Workaround a timing issue when using the "Help: Logtalk Handbook" command for the first time that may cause the Handbook to not open
 
 ## [0.83.1]